feat(wysiwyg): allow lazy loading braft editor to reduce initial payload
